Bulk commit working quiz *happy noises*

+++ Backend/Data/Models/Highscore.cs
@@ -12,9 +12,9 @@ namespace Backend.Data.Models
         public string Name { get; set; }
         public DateTime TimeStamp { get; set; }
         public int PointsAchieved { get; set; }
-        public long Duration { get; set; }
+        public int Duration { get; set; }
 
-        public float PointsWeighted { get; private set; }
+        public int PointsWeighted { get; private set; }
 
         public ICollection<Category> Categories { get; set; } = new HashSet<Category>();
     }
